LayoutTests/imported/w3c/web-platform-tests/import-maps/csp/hash-failure-expected.txt

CONSOLE MESSAGE: Refused to execute a script because its hash, its nonce, or 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.

PASS Importmap should not be accepted due to wrong hash
